requirements.txt
aioeasywebdav==2.4.0
aiohttp==2.3.7
alabaster==0.7.10
amqp==2.2.2
appdirs==1.4.3
asn1crypto==0.24.0
async-timeout==2.0.0
aws-xray-sdk==0.95
Babel==2.5.1
bcrypt==3.1.4
beautifulsoup4==4.6.0
billiard==3.5.0.3
biopython==1.70
boto==2.48.0
boto3==1.5.8
botocore==1.8.22
cachetools==2.0.1
celery==4.1.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
ConfigArgParse==0.12.0
configparser==3.5.0
confusable-homoglyphs==2.0.2
cookies==2.2.1
coreapi==2.3.3
coreschema==0.0.4
crc16==0.1.1
cryptography==2.1.4
dill==0.2.7.1
Django==2.0.1
django-celery-beat==1.1.0
django-cors-headers==2.1.0
django-cprofile-middleware==1.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-mysql==2.2.0
django-registration==2.3
django-rest-framework==0.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docker==2.7.0
docker-pycreds==0.2.1
docutils==0.14
dropbox==8.5.1
enum-compat==0.0.2
ephem==3.7.6.0
eventlet==0.21.0
Flask==0.12.2
flower==0.9.2
ftputil==3.4
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gevent==1.2.2
google==1.9.3
google-api-core==0.1.3
google-auth==1.2.1
google-cloud==0.32.0
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.0
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.0
google-cloud-core==0.28.0
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.0
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.0
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.0
google-cloud-runtimeconfig==0.28.0
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.0
google-cloud-videointelligence==1.0.0
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
greenlet==0.4.12
grpc-google-iam-v1==0.11.4
grpcio==1.8.3
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsondiff==1.1.1
jsonpickle==0.9.5
kombu==4.1.0
lockfile==0.12.2
lxml==4.1.1
MarkupSafe==1.0
mock==2.0.0
moto==1.1.25
multidict==3.3.2
mysqlclient==1.3.12
numpy==1.13.3
oauth2client==3.0.0
openapi-codec==1.3.2
paramiko==2.4.0
pbr==3.1.1
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.1
psutil==5.4.3
pyaml==17.12.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.3
PyNaCl==1.2.1
pysftp==0.2.9
python-daemon==2.1.2
python-dateutil==2.6.1
python-redis==0.1.7
pytz==2017.3
PyYAML==3.12
ratelimiter==1.2.0.post0
redis==2.10.6
requests==2.18.4
requests-mock==1.4.0
rsa==3.4.2
s3transfer==0.1.12
setproctitle==1.1.10
setuptools-scm==1.15.6
simplejson==3.13.2
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
tornado==4.5.2
uritemplate==3.0.0
urllib3==1.22
uuid==1.30
vine==1.1.4
websocket-client==0.46.0
Werkzeug==0.14.1
wrapt==1.10.11
xmltodict==0.11.0
yarl==0.17.0
